Start on CreepInfo class creation
