+++ src/types.d.ts
@@ -44,80 +44,6 @@ interface ScreepsMultimeterWatch {
   values?: { [index: string]: any };
 }
 
-interface CreepMemory {
-  role: CreepRole;
-  task: CreepTask;
-  // Undefined if the creep is spawning
-  room: string;
-  /** A source assigned to this creep by id */
-  assignedSource?: Id<Source> | undefined;
-  /** A construction site assigned to this creep by id */
-  assignedConstruction?: string | undefined;
-  /** A structuring needing repairs that this creep is repairing */
-  assignedRepairs?: Id<Structure> | undefined;
-  /** A spot assigned to this creep */
-  spot?: RoomPosition | undefined;
-  /** Whether to prevent this creep from being renewed */
-  noRenew?: boolean | undefined;
-  /** The room this claimer creep is targetting */
-  claimTarget?: string | undefined;
-  /** A path for this creep to use, serialized with `Room.serializePath` */
-  path?: string;
-  /** The room the path originated in, to know when to recreate path */
-  pathStartRoom?: string;
-  /** The room this creep is targetting */
-  roomTarget?: string;
-  /** Whether this creep should have attack notifications enabled */
-  attackNotifications?: boolean;
-}
-
-// The exact task depends also on the role
-declare const enum CreepTask {
-  /** Role indicating the creep is freshly spawned (i.e. uninit) */
-  fresh = "fresh",
-  /** Task indicating the creep is waiting for further instructions/state change */
-  idle = "idle",
-  harvest = "harvest",
-  deposit = "deposit",
-  getEnergy = "get_energy",
-  build = "build",
-  repair = "repair",
-  renew = "renew",
-  claim = "claim",
-  reserve = "reserve",
-  /** Move to target */
-  move = "move",
-  /** Attack hostiles */
-  attack = "attack",
-  /** Scout rooms without necessarily staying and signing */
-  scout = "scout",
-}
-
-declare const enum CreepRole {
-  /** Simple creep that performs the harvest and deposit actions */
-  harvester = "harvester",
-  /** Creep that mines into a container near to the source */
-  miner = "miner",
-  /** Creep that constructs buildings */
-  builder = "builder",
-  /** Creep that gets energy and deposits energy to spawn then controller */
-  upgrader = "upgrader",
-  /** Creep that hauls energy between sources and deposits energy */
-  hauler = "hauler",
-  /** Creep that moves to other rooms to claim them */
-  claimer = "claimer",
-  /** Creep that keeps energy in spawns, extensions, and towers */
-  tender = "tender",
-  /** Creep that works mineral deposits */
-  extractor = "extractor",
-  /** Creep that hauls between remotes */
-  remoteHauler = "remote_hauler",
-  /** Creep that moves to a room to provide vision */
-  scout = "scout",
-  /** Creep that guards rooms and their remotes */
-  guard = "guard",
-}
-
 declare const enum InfoType {
   general = "general",
   spawn = "spawn",
